// src/store/modules/user/types.ts
export interface UserState {
  createBy?: string;
  createTime?: string;
  updateBy?: string;
  updateTime?: string;
  userId?: string;
  username?: string;
  realName?: string;
  sex?: string;
  birthday?: string;
  phone?: string;
  email?: string;
  avatar?: string;
  avatarURL?: string;
  employeeCode?: string;
  accountId?: string;
  loginCount?: string;
  todayLoginCount?: string;
  lastLoginTime?: string;
  lockFlag?: string;
  delFlag?: string;
  sort?: string;
  description?: string;
}